Web audio partially works
Beeps work, buffer makes noises, need to figure out how to correctly make the buffer before web audio fully works.

+++ crates/c8_device/src/message.rs
@@ -17,7 +17,7 @@ pub enum DeviceMessage {
     Beep(u8),
 
     /// Set the pitch of the beeper
-    SetPitch(f64),
+    SetPitch(f32),
 
     /// New audio buffer created
     NewAudioBuffer,
